in sirisha as of 27/08 16:30 with lot of errors

+++ commons/src/main/java/be/fooda/backend/commons/model/template/store/request/FoodaStoreDeliveryLocationsItemReq.java
@@ -2,13 +2,14 @@ package be.fooda.backend.commons.model.template.store.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @Builder
 public class FoodaStoreDeliveryLocationsItemReq {
     private String country;
-    private Double deliveryCost;
+    private BigDecimal deliveryCost;
     private String city;
     private String municipality;
     private LocalDateTime deliveryTime;
